+++ lib/WebImage/ServiceManager/IFactory.php
@@ -1,5 +1,9 @@
 <?php
 
+namespace WebImage\ServiceManager;
+
+use WebImage\ServiceManager\IServiceManager;
+
 interface IFactory {
 	public function createService(IServiceManager $sm);
 }