+++ lib/WebImage/ServiceManager/IServiceManagerConfig.php
@@ -1,4 +1,9 @@
 <?php
+
+namespace WebImage\ServiceManager;
+
+use WebImage\ServiceManager\IServiceManager;
+
 interface IServiceManagerConfig {
 	public function configureServiceManager(ServiceManager $serviceManager);
 }